+++ CHANGELOG.md
@@ -15,7 +15,9 @@
 - Added ability to reuse built-in styles (scroll,button['submit'] etc.). See [FAQ/sharedStyles](https://yegorich555.github.io/web-ui-pack/faq#shared-styles)
 - [SelectManyControl](src/controls/selectMany.ts) [**demo**](https://yegorich555.github.io/web-ui-pack/control/selectMany). Added ability to sort/order items via drag&drop OR Arrows + Shift
 - Added helper [**animateStack**](src/helpers/animate.ts)
-- Added option **animation: Animations.stack** for [PopupElement](src/popup/popupElement.ts)
+- [PopupElement](src/popup/popupElement.ts)
+  - Added option **animation: Animations.stack** for [PopupElement](src/popup/popupElement.ts)
+  - Extended attribute **placement**. Now it has predefined opposite rules (so `right-middle` means `[right-middle, left-middle]`)
 
 ## 0.5.2 (Mar 29, 2023)